apis for greivance
