+++ grievance-management-service/src/main/java/com/scaler/grievancemanagementservice/controllers/GreivanceController.java
@@ -1,14 +1,13 @@
-package controllers;
+package com.scaler.grievancemanagementservice.controllers;
 
-import dtos.GreivanceDto;
-import entities.Greivance;
+import com.scaler.grievancemanagementservice.entities.Greivance;
+import com.scaler.grievancemanagementservice.services.GrevienceService;
+import com.scaler.grievancemanagementservice.dtos.GreivanceDto;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import services.GrevienceService;
-import services.SelfGreivanceService;
 
 import java.util.List;
 import java.util.Optional;
@@ -31,6 +30,7 @@ public class GreivanceController implements GreivanceOperations{
 
     @Override
     public GreivanceDto createGreivance(@RequestBody GreivanceDto greivanceDto) {
+        System.out.println("this is post mapping...");
         return grevienceService.createGreivance(greivanceDto);
     }
