--- grievance-management-service/src/main/java/dtos/GreivanceDto.java
@@ -1,17 +1,0 @@
-package dtos;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class GreivanceDto {
-    private String description;
-    private String assignee;
-    private String createdBy;
-    private String updatedBy;
-    private Long timeUpdated;
-    private Long timeCreated;
-}
