--- grievance-management-service/src/main/java/entities/Role.java
@@ -1,21 +1,0 @@
-package entities;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-
-@Entity
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class Role {
-    @Id
-    @GeneratedValue
-    private int role_id;
-    private String role_name;
-    private long time_created;
-}
