+++ grievance-management-service/src/main/resources/application.properties
@@ -1,7 +1,9 @@
 server.port=8082
 
-spring.datasource.url=jdbc:h2:mem:testdb
-spring.datasource.driverClassName=org.h2.Driver
-spring.datasource.username=sa
-spring.datasource.password=password
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+spring.jpa.hibernate.ddl-auto=update
+spring.datasource.url=jdbc:mysql://scalergrievancessystem-do-user-14974559-0.c.db.ondigitalocean.com:25060/defaultdb
+spring.datasource.username=doadmin
+spring.datasource.password=AVNS_nqgkYcfYzE3hZxMrilG
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+spring.datasource.hibernate.dialect=org.hibernate.dialect.MySQLDialect
+spring.jpa.show-sql=true
